Migrate CallList to new list component
